tests: use the shared selector helpers across the tests
