+++ tests/queryEditor-avro.spec.ts
@@ -9,6 +9,7 @@ import {
   setTableVisualization,
   openPartitionSelector,
   selectAllPartitionsOption,
+  selectMessageFormat,
 } from './test-utils';
 
 interface AvroProducerOptions {
@@ -91,44 +92,8 @@ function startTransactionalAvroKafkaProducer(topic: string): {
   });
 }
 
-async function findMessageFormatSelector(page: Page): Promise<Locator | null> {
-  const messageFormatApproaches = [
-    page
-      .locator('div')
-      .filter({ hasText: /^JSON$/ })
-      .nth(2), // The intercepting parent element
-    page.getByText('JSON').locator('../..'), // Go up two levels to find clickable parent
-    page.locator('.css-1eu65zc').filter({ hasText: /JSON/ }), // Direct selection of intercepting element
-    page.getByText('JSON').filter({ hasText: /^JSON$/ }), // Original approach
-    page.locator('button').filter({ hasText: /^JSON$/ }), // Button with exact JSON text
-    page.getByText('Message Format').locator('..').locator('button').first(), // Button near Message Format label
-    page.locator('[data-testid*="select"]'), // Any element with select in testid
-  ];
-
-  for (let i = 0; i < messageFormatApproaches.length; i++) {
-    const approach = messageFormatApproaches[i];
-    if (await approach.isVisible({ timeout: 1000 })) {
-      console.log(`Message format selector found using approach ${i}: ${approach.toString()}`);
-      return approach;
-    }
-  }
-  return null;
-}
-
 async function selectAvroMessageFormat(page: Page): Promise<void> {
-  // Wait for the page to stabilize after filling topic name
-  await expect(page.getByText('Message Format')).toBeVisible({ timeout: 5000 });
-
-  console.log('Looking for Message Format selector among buttons...');
-
-  const foundSelector = await findMessageFormatSelector(page);
-
-  // Message format selector MUST be found for Avro tests
-  expect(foundSelector).not.toBeNull();
-
-  // Click the message format selector
-  await foundSelector!.first().click();
-  await page.getByText('Avro').click();
+  await selectMessageFormat(page, 'Avro');
 }
 
 function getAvroSchemaSourceLocator(page: Page): Locator {
