+++ tests/queryEditor-lineprotocol.spec.ts
@@ -1,10 +1,11 @@
 /// <reference types="node" />
 
 import { test, expect } from '@grafana/plugin-e2e';
-import { Locator, Page } from '@playwright/test';
 import { ChildProcess, spawn } from 'child_process';
 import { accessSync, constants } from 'fs';
 import {
+  selectMessageFormat,
+  selectTimestampPrecision,
   openPartitionSelector,
   selectAllPartitionsOption,
   setTableVisualization,
@@ -63,83 +64,6 @@ function startLineProtocolKafkaProducer(topic: string): ProducerHandle {
   });
 
   return { producer, exitPromise };
-}
-
-async function findMessageFormatSelector(page: Page): Promise<Locator | null> {
-  const messageFormatApproaches = [
-    page
-      .locator('div')
-      .filter({ hasText: /^JSON$/ })
-      .nth(2),
-    page.getByText('JSON').locator('../..'),
-    page.locator('.css-1eu65zc').filter({ hasText: /JSON/ }),
-    page.getByText('JSON').filter({ hasText: /^JSON$/ }),
-    page.locator('button').filter({ hasText: /^JSON$/ }),
-    page.getByText('Message Format').locator('..').locator('button').first(),
-    page.locator('[data-testid*="select"]'),
-  ];
-
-  for (const approach of messageFormatApproaches) {
-    if (await approach.isVisible({ timeout: 1000 })) {
-      return approach;
-    }
-  }
-
-  return null;
-}
-
-async function selectMessageFormat(page: Page, optionName: string): Promise<void> {
-  await expect(page.getByText('Message Format')).toBeVisible({ timeout: 5000 });
-
-  const foundSelector = await findMessageFormatSelector(page);
-  expect(foundSelector).not.toBeNull();
-
-  await foundSelector!.first().click();
-
-  const option = page
-    .getByLabel('Select options menu')
-    .getByText(optionName)
-    .or(page.getByRole('option', { name: optionName }));
-  await expect(option.first()).toBeVisible({ timeout: 5000 });
-  await option.first().click();
-}
-
-async function findTimestampPrecisionSelector(page: Page): Promise<Locator | null> {
-  const selectorApproaches = [
-    page.getByText('Timestamp Precision').locator('..').locator('button').first(),
-    page
-      .getByRole('combobox')
-      .filter({ hasText: /Auto-detect|Nanoseconds|Microseconds|Milliseconds|Seconds/ })
-      .first(),
-    page
-      .locator('div')
-      .filter({ hasText: /^Auto-detect$/ })
-      .first(),
-  ];
-
-  for (const approach of selectorApproaches) {
-    if (await approach.isVisible({ timeout: 1000 })) {
-      return approach;
-    }
-  }
-
-  return null;
-}
-
-async function selectTimestampPrecision(page: Page, optionName: string): Promise<void> {
-  await expect(page.getByText('Timestamp Precision')).toBeVisible({ timeout: 5000 });
-
-  const selector = await findTimestampPrecisionSelector(page);
-  expect(selector).not.toBeNull();
-
-  await selector!.first().click();
-
-  const option = page
-    .getByLabel('Select options menu')
-    .getByText(optionName)
-    .or(page.getByRole('option', { name: optionName }));
-  await expect(option.first()).toBeVisible({ timeout: 5000 });
-  await option.first().click();
 }
 
 test.describe.serial('Kafka Query Editor - Line Protocol Tests', () => {
